1. Add closure table & backend api to allow admin to set a space/accommodation temporary unavailable (any time without booking rule restriction) 2. Add event table as an abstraction layer between booking & closure to availability 3. Add event api that allow get list of event for a space/accommodation in given period 4. Make event service to update availability when there is change in event for a space/accommodation 5. Refactor booking module to move booking table to each belongs module(space/service/accommodation) 6. Remove redundant export excel api & file

+++ umpoint/umpoint-application/src/main/java/my/edu/um/umpoint/modules/space/dto/SpcAvailabilityDTO.java
@@ -7,8 +7,6 @@ import lombok.Data;
 import java.io.Serial;
 import java.io.Serializable;
 
-import java.math.BigDecimal;
-
 /**
  * Space Availability
  *
@@ -28,7 +26,7 @@ public class SpcAvailabilityDTO implements Serializable {
 	private Long spaceId;
 
 	@SchemaProperty(name = "Year")
-	private BigDecimal year;
+	private Integer year;
 
 	@SchemaProperty(name = "Availability of space, consist of 366*24 bit, 1 represent available in specific hour in one year")
 	private byte[] availability;
